beater/config: treat null agent config as empty
