+++ changelogs/head.asciidoc
@@ -18,6 +18,7 @@ https://github.com/elastic/apm-server/compare/8.5\...main[View commits]
 ==== Bug fixes
 - Set `error.id` for OpenTelemetry exception span events {pull}9372[9372]
 - Add an `apm-server` user to /etc/passwd in the Docker images {pull}9496[9496]
+- Null APM agent configuration entries are now treated as empty, and will not cause the server to exit with an error {pull}9546[9546]
 
 [float]
 ==== Intake API Changes
